# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Claudia Vanegas POS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    put:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── PRODUCTS ────────────────────────────────────────────────────────────────
...
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: List all suppliers
      responses:
        "200":
          description: List of suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create a supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    get:
      operationId: getSupplier
      tags: [suppliers]
      summary: Get a supplier by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
        "404":
          description: Not found
    put:
      operationId: updateSupplier
      tags: [suppliers]
      summary: Update a supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    delete:
      operationId: deleteSupplier
      tags: [suppliers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── PURCHASE ORDERS ─────────────────────────────────────────────────────────
...
  /sales:
    get:
      operationId: listSales
      tags: [sales]
      summary: List all sales/invoices
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
        - name: paymentType
          in: query
          schema:
            type: string
            enum: [contado, credito]
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
    post:
      operationId: createSale
      tags: [sales]
      summary: Create a sale (checkout from POS)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaleInput"
      responses:
        "201":
          description: Sale created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
  /sales/{id}:
    get:
      operationId: getSale
      tags: [sales]
      summary: Get a sale by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
        "404":
          description: Not found

  # ─── ACCOUNTS RECEIVABLE ─────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, cajero]
        createdAt:
          type: string
          format: date-time
    UserInput:
      type: object
      required: [email, name, role, password]
      properties:
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, cajero]
        password:
          type: string
    UserUpdate:
      type: object
      properties:
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, cajero]
        password:
          type: string

    # ─── PRODUCT ─────────────────────────────────────────────────────────────
...
    Supplier:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        contact:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    SupplierInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        contact:
          type: string
        email:
          type: string
        phone:
          type: string
    SupplierUpdate:
      type: object
      properties:
        name:
          type: string
        contact:
          type: string
        email:
          type: string
        phone:
          type: string

    # ─── PURCHASE ORDER ──────────────────────────────────────────────────────
...
    SaleItem:
      type: object
      required: [id, productId, productName, qty, unitPrice, subtotal]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        qty:
          type: integer
        unitPrice:
          type: number
        subtotal:
          type: number
    SaleItemInput:
      type: object
      required: [productId, qty, unitPrice]
      properties:
        productId:
          type: integer
        qty:
          type: integer
        unitPrice:
          type: number
    Sale:
      type: object
      required: [id, userId, paymentType, total, items, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: ["string", "null"]
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        paymentType:
          type: string
          enum: [contado, credito]
        total:
          type: number
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
        createdAt:
          type: string
          format: date-time
    SaleInput:
      type: object
      required: [paymentType, items]
      properties:
        customerId:
          type: integer
        paymentType:
          type: string
          enum: [contado, credito]
        notes:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItemInput"

    # ─── ACCOUNTS RECEIVABLE ─────────────────────────────────────────────────
...