# lib/api-zod/src/generated/types/supplierUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Claudia Vanegas POS API
 * OpenAPI spec version: 0.1.0
 */

export interface SupplierUpdate {
  name?: string;
  contact?: string;
  email?: string;
  phone?: string;
}